+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.1
+
+1. StatefulField support `onInitialed` , used to listen `FormeFieldController` initialed
+2. add `FormeValidateUtils`
+3. bug fix : onErrorChanged and errorTextListenable not triggered in build 
+
 ## 2.0.0
 
 forme is completely rewrite version of https://pub.dev/packages/form_builder much more powerful and won't break your layout